Добавил загрузку аватарки с помощью Kingfisher
